+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Fix compute-plan-less compute tasks all using the same lock key (#205)
+- Fix asset buffer skipping some data samples (#228)
 
 ## [0.3.0] - 2021-08-17
 ### Added